Refactoring and some documentation
